refactor(teamSelection): extract grid + sidebar view components
Move the club grid (search, grouped team cards, InfoStat) and the
selected-team sidebar (DetailTile, key players, competitions) into
TeamSelectionGrid and TeamSelectionSidebar. getReputationLabel/InfoStat
(grid-only) and DetailTile (sidebar-only) move with their single
consumer; getTeamAvgOvr / player-count stay in the parent (shared) and
pass as props. Adds a render smoke test (TeamSelection had no coverage),
exercising the assembled page through the real component tree.

TeamSelection.tsx: 825 -> 564 lines.
